Fix signature helper
The idea is that if we already are **in** a call expression then we want the
signature of this call expression.
If we aren't then we want to signature of the call expression we are on.
If we are neither in or on an call expression, then we can't resolve any signature.

The idea is that 3 scenarios:
1. The cursor is inside a call expression (nested or not)
2. The cursor is on a call expression
3. The cursor is not inside or on a call expression

My idea was to focus more on the `ArgumentExpressionList` than on the
call expression.
There is two different kind of call expression which makes it awkward,
but for any arguements list expression the parent will always gave us
the closest call expression.
Also, in order ot fix the problem of comma being part of string which
messed up the active argument, I wanted to rely on the number of
`ArgumentExpression` instead on the comma characted.

It should be trivial:
1. Find the closest call expression (first ancestor in the AST)
2. Don't care about the `ArgumentExpressionList` since we are not in it
3. throw an exception

But the AST does not work like that, for example:
* If the cursor is on a variable which is an argument, the node is
`VariableExpression` so we need to find the closest
`ArgumentExpressionList`
* If we are on a blank line (multiline arguments) then the node at the
position is resolve to the call expression instead of the argument list,
so we need to check if there is a child `ArgumentExpressionList`

At then end it's messy and I'm not sure of every possible cases so I
decided to trust our tests.
I wrote every scenarios I could think of and that was not already in the
test file.