remove the need for WordAtOffset
It was awkard to add the extraction of the annotation there.
We were looking for a qualified name, which was already possible, but we
also need to makee sure there is a @ at the begening of it.
Since it feels a bit like a niche use case I prefer to do it directly in
the completor for now.
Rule of three, we will see later if we really need to be able to resuse
it.