add suggestions instead of replacing
Before all I did was trying to find the most accurate name to use
depending on the import table.
It actually not the job of the completor but the responsibility of the
user.

This commit will instead resolve a class suggestion to a list of
suggestions in order for the user to pick the one he wants.