extract the logic into a trait
I want to extract the logic in order to reuse it in
DoctrineAnnotationCompletor
Inheritance is not suited in this case, the only remaining options I can
think of are:
* Using a static method in an helper class
* Injecting a new dependency

I don't think injecting a dependency is a good solution here since I
can't think of a different implementation I might want to inject.
The helper class with a static method is a possibility but I think trait
are also meant for scenarios like this one.

So I kept the trait to open a new debate :)