WorseNamedParameterCompletor: ignored when completing a variable
I often find myself in the situation where a function/method's parameter
name match the name of a local variable I want to provide as argument.

In this situation the completion results include the named parameter
which can be anoying depending on the sort done by the client side.

I think it's sane to not include named parameters when we already
started to type a variable name, our little `$` in php will finally pay
off :)